requirements.txt
addict==2.4.0
albumentations==1.1.0
dgl==0.4.3.post2
einops==0.4.0
h5py==3.6.0
histocartography==0.2.1
imagecodecs==2021.11.20
mmcv-full==1.5.0
omegaconf==2.2.0
opencv-python==4.1.2.30
opencv-python-headless==4.5.5.62
openslide-python==1.1.2
Pillow==8.4.0
protobuf==3.19.1
pytorch-lightning==1.5.10
pytorch-toolbelt==0.4.4
PyYAML==6.0
scikit-image==0.17.2
scikit-learn==1.0.2
tensorboard==2.7.0
tifffile==2021.11.2
timm==0.5.4
torchmetrics==0.6.2
tqdm==4.62.3
wandb==0.12.10
loguru

